% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting_utils.R
\name{plot_distribution_by_class}
\alias{plot_distribution_by_class}
\title{Plotting Output Distribution By Class}
\usage{
plot_distribution_by_class(analysis_object)
}
\arguments{
\item{analysis_object}{Fitted analysis_object with 'fine_tuning()'.}
}
\value{
analysis_object
}
\description{
Plotting Output Distribution By Class
}
