% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting_utils.R
\name{plot_pr_curve}
\alias{plot_pr_curve}
\title{Plotting Precision-Recall Curve}
\usage{
plot_pr_curve(analysis_object)
}
\arguments{
\item{analysis_object}{Fitted analysis_object with 'fine_tuning()'.}
}
\value{
analysis_object
}
\description{
Plotting Precision-Recall Curve
}
