% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting_utils.R
\name{plot_scatter_predicitions}
\alias{plot_scatter_predicitions}
\title{Plotting Observed vs Predictions}
\usage{
plot_scatter_predicitions(analysis_object)
}
\arguments{
\item{analysis_object}{Fitted analysis_object with 'fine_tuning()'.}
}
\value{
analysis_object
}
\description{
Plotting Observed vs Predictions
}
