% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting_utils.R
\name{plot_scatter_predictions}
\alias{plot_scatter_predictions}
\title{Plotting Observed vs Predictions}
\usage{
plot_scatter_predictions(analysis_object)
}
\arguments{
\item{analysis_object}{Fitted analysis_object with 'fine_tuning()'.}
}
\value{
analysis_object
}
\description{
The \strong{plot_scatter_predictions()} function generates scatter plots between observed and predicted values,
providing direct visual assessment of model predictive accuracy. The function displays both training and
test results side by side, enabling evaluation of model generalization performance and identification of
potential overfitting.
}
\examples{
# Note: For obtaining the observed vs. predicted values plot the user needs to
# complete till fine_tuning( ) function of the MLwrap pipeline.

\dontrun{

wrap_object <- preprocessing(df = sim_data,
                             formula = psych_well ~ depression + emot_intel + resilience,
                             task = "regression")
wrap_object <- build_model(wrap_object, "Random Forest")
wrap_object <- fine_tuning(wrap_object, "Grid Search CV")

# And then, you can obtain the observed vs predicted values plot.

plot_scatter_predictions(wrap_object)

}

}
