% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocessing.R
\name{preprocessing}
\alias{preprocessing}
\title{Preprocessing Data Matrix}
\usage{
preprocessing(
  df,
  formula,
  task = "regression",
  num_vars = NULL,
  cat_vars = NULL,
  norm_num_vars = "all",
  encode_cat_vars = "all",
  y_levels = NULL
)
}
\arguments{
\item{df}{Input Dataframe. Either a data.frame or tibble.}

\item{formula}{Modelling Formula. A string of characters or formula.}

\item{task}{Modelling Task. Either "regression" or "classification".}

\item{num_vars}{Optional vector of names of the numerical features.}

\item{cat_vars}{Optional vector of names of the categorical features.}

\item{norm_num_vars}{Normalize numeric features as z-scores. Either vector of
names of numerical features to be normalized or "all" (default).}

\item{encode_cat_vars}{One Hot Encode Categaorical Features. Either vector of
names of categorical features to be encoded or "all" (default).}

\item{y_levels}{Optional ordered vector with names of the target variable levels.}
}
\value{
A tidy_object
}
\description{
Preprocessing Data Matrix
}
