% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/InterpretableML.R
\name{sensitivity_analysis}
\alias{sensitivity_analysis}
\title{Perform Sensitivity Analysis and Interpretable ML methods}
\usage{
sensitivity_analysis(tidy_object, type = "PFI", metric = NULL)
}
\arguments{
\item{tidy_object}{Tidy_Object created from fine_tuning function.}

\item{type}{Type of method used. A string of the method name: "PFI" (Permutation Feature Importance),
"SHAP" (SHapley Additive exPlanations), "Integrated Gradients" (Neural Network only) or
"Olden" (Neural Network only).}

\item{metric}{Metric used for "PFI" method (Permutation Feature Importance).
A string of the name of metric (see Metrics).}
}
\value{
Updated tidy_object
}
\description{
Perform Sensitivity Analysis and Interpretable ML methods
}
