% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/results.R
\name{show_results}
\alias{show_results}
\title{Showcase Summary Results and Plots}
\usage{
show_results(
  tidy_object,
  summary = FALSE,
  roc_curve = FALSE,
  pr_curve = FALSE,
  gain_curve = FALSE,
  lift_curve = FALSE,
  dist_by_class = FALSE,
  reliability_plot = FALSE,
  confusion_matrix = FALSE,
  scatter_residuals = FALSE,
  scatter_predictions = FALSE,
  residuals_dist = FALSE,
  new_data = "test"
)
}
\arguments{
\item{tidy_object}{Tidy_Object created from fine_tuning function.}

\item{summary}{Whether to plot summary results table. Boolean (FALSE by default).}

\item{roc_curve}{Whether to plot ROC Curve (Classification task only). Boolean (FALSE by default).}

\item{pr_curve}{Whether to plot ROC Curve (Classification task only). Boolean (FALSE by default).}

\item{gain_curve}{Whether to plot ROC Curve (Classification task only). Boolean (FALSE by default).}

\item{lift_curve}{Whether to plot ROC Curve (Classification task only). Boolean (FALSE by default).}

\item{dist_by_class}{Whether to plot distribution of output probability by class (Classification task only).
Boolean (FALSE by default).}

\item{reliability_plot}{Whether to plot Reliability Plot (Binary Classification task only). Boolean (FALSE by default).}

\item{confusion_matrix}{Whether to Confusion Matrix (Classification task only). Boolean (FALSE by default).}

\item{scatter_residuals}{Whether to plot Residuals vs Predictions (Regression task only). Boolean (FALSE by default).}

\item{scatter_predictions}{Whether to plot Predictions vs Observed (Regression task only). Boolean (FALSE by defaut).}

\item{residuals_dist}{Whether to plot Residuals Distribution (Regression task only). Boolean (FALSE by default).}

\item{new_data}{Data to be used for Confusion Matrix, Reliability Plot, Distribution by Class Plot,
Residuals vs Predictions Plot, Predictions vs Observed Plot and Residuals Distribution Plot.
A string with the name of the data_set: "train", "validation", "test" (default) or "all".}
}
\value{
Updated tidy_object
}
\description{
Showcase Summary Results and Plots
}
