% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_data.R
\docType{data}
\name{sim_data}
\alias{sim_data}
\title{Example Data Set}
\format{
A data frame with X rows and Y columns:
\describe{
\item{variable1}{Description}
\item{variable2}{Description}
}
}
\source{
Original file converted from your_file.xlsx
}
\usage{
sim_data
}
\description{
This dataset contains XYZ...
}
\keyword{datasets}
