% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table_utils.R
\name{table_best_hyperparameters}
\alias{table_best_hyperparameters}
\title{Best Hyperparameter Configuration}
\usage{
table_best_hyperparameters(analysis_object, show_table = FALSE)
}
\arguments{
\item{analysis_object}{Fitted analysis_object with 'fine_tuning()'.}

\item{show_table}{Boolean. Whether to print the table.}
}
\value{
Tibble with best hyperparameter configuration.
}
\description{
Best Hyperparameter Configuration
}
