% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table_utils.R
\name{table_olden_results}
\alias{table_olden_results}
\title{Olden Results Table}
\usage{
table_olden_results(analysis_object, show_table = FALSE)
}
\arguments{
\item{analysis_object}{Fitted analysis_object with 'sensitivity_analysis(methods = "Olden")'.}

\item{show_table}{Boolean. Whether to print the table.}
}
\value{
Tibble or list of tibbles (multiclass classification) with Olden results.
}
\description{
Olden Results Table
}
