% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table_utils.R
\name{table_pfi_results}
\alias{table_pfi_results}
\title{Permutation Feature Importance Results Table}
\usage{
table_pfi_results(analysis_object, show_table = FALSE)
}
\arguments{
\item{analysis_object}{Fitted analysis_object with 'sensitivity_analysis(methods = "PFI")'.}

\item{show_table}{Boolean. Whether to print the table.}
}
\value{
Tibble or list of tibbles (multiclass classification) with PFI results.
}
\description{
Permutation Feature Importance Results Table
}
