% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_creation.R
\name{build_model}
\alias{build_model}
\title{Create ML Model}
\usage{
build_model(tidy_object, model_names, hyperparameters = NULL)
}
\arguments{
\item{tidy_object}{Tidy_Object created from preprocessing function.}

\item{model_names}{Name of the ML Model. A string of the model name: "Neural Network",
"Random Forest", "SVM" or "XGBOOST".}

\item{hyperparameters}{Hyperparameters of the ML model. List containing the name of the hyperparameter
and its value or range of values.}
}
\value{
Updated tidy_object
}
\description{
Create ML Model
}
