% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_tuning.R
\name{fine_tuning}
\alias{fine_tuning}
\title{Fine Tune ML Model}
\usage{
fine_tuning(tidy_object, tuner, metrics, plot_results = F, verbose = FALSE)
}
\arguments{
\item{tidy_object}{Tidy_Object created from build_model function.}

\item{tuner}{Name of the Hyperparameter Tuner. A string of the tuner name: "Bayesian Optimization" or
"Grid Search CV".}

\item{metrics}{Metric used for Model Selection. A string of the name of metric (see metrics).}

\item{plot_results}{Whether to plot the tuning results. Boolean TRUE or FALSE (default).}

\item{verbose}{Whether to show tuning process. Boolean TRUE or FALSE (default).}
}
\value{
Updated tidy_object
}
\description{
Fine Tune ML Model
}
