% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_tuning.R
\name{fine_tuning}
\alias{fine_tuning}
\title{Fine Tune ML Model}
\usage{
fine_tuning(analysis_object, tuner, metrics, plot_results = F, verbose = FALSE)
}
\arguments{
\item{analysis_object}{analysis_object created from build_model function.}

\item{tuner}{Name of the Hyperparameter Tuner. A string of the tuner name: "Bayesian Optimization" or
"Grid Search CV".}

\item{metrics}{Metric used for Model Selection. A string of the name of metric (see Metrics).}

\item{plot_results}{Whether to plot the tuning results. Boolean TRUE or FALSE (default).}

\item{verbose}{Whether to show tuning process. Boolean TRUE or FALSE (default).}
}
\value{
Updated analysis_object
}
\description{
Fine Tune ML Model
}
\section{Tuners}{

\subsection{Bayesian Optimization}{
\itemize{
\item Initial data points: 20
\item Maximum number of iterations: 25
\item Convergence after 5 iterations without improvement
\item Train / Validation / Test : 0.6 / 0.2 / 0.2
}
}

\subsection{Grid Search CV}{
\itemize{
\item Number of Folds: 5
\item Max grid size per hyperparameter: 10
\item Train / Test : 0.75 / 0.25
}
}
}

\section{Metrics}{

\subsection{Regression Metrics}{
\itemize{
\item rmse
\item mae
\item mpe
\item mape
\item ccc
\item smape
\item rpiq
\item rsq
}
}

\subsection{Classification Metrics}{
\itemize{
\item accuracy
\item bal_accuracy
\item recall
\item sensitivity
\item specificity
\item kap
\item f_meas
\item mcc
\item j_index
\item detection_prevelance
\item roc_auc
\item pr_auc
\item gain_capture
\item brier_class
\item roc_aunp
}
}
}

