% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting_utils.R
\name{plot_calibration_curve}
\alias{plot_calibration_curve}
\title{Plotting Calibration Curve}
\usage{
plot_calibration_curve(analysis_object)
}
\arguments{
\item{analysis_object}{Fitted analysis_object with 'fine_tuning()'.}
}
\value{
analysis_object
}
\description{
Binary Classification Only
}
