% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting_utils.R
\name{plot_confusion_matrix}
\alias{plot_confusion_matrix}
\title{Plotting Confusion Matrix}
\usage{
plot_confusion_matrix(analysis_object)
}
\arguments{
\item{analysis_object}{Fitted analysis_object with 'fine_tuning()'.}
}
\value{
analysis_object
}
\description{
Plotting Confusion Matrix
}
