% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting_utils.R
\name{plot_gain_curve}
\alias{plot_gain_curve}
\title{Plotting Gain Curve}
\usage{
plot_gain_curve(analysis_object)
}
\arguments{
\item{analysis_object}{Fitted analysis_object with 'fine_tuning()'.}
}
\value{
analysis_object
}
\description{
The \strong{plot_gain_curve()} plots cumulative gain as a function of sorted
population percentile when observations are ranked by descending predicted
probability. For each percentile threshold, calculates the ratio of positive
class proportion in the top-ranked subset relative to overall positive class
proportion, quantifying model's efficiency in concentrating target cases at
the top of rankings.
}
\examples{
# Note: For obtaining the gain curve plot the user needs to complete till
# fine_tuning( ) function of the MLwrap pipeline and only with categorical
# outcome.
# See the full pipeline example under plot_calibration_curve()
# Final call signature:
# plot_gain_curve(wrap_object)
}
\seealso{
\code{\link{plot_calibration_curve}}
}
