% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting_utils.R
\name{plot_gain_curve}
\alias{plot_gain_curve}
\title{Plotting Gain Curve}
\usage{
plot_gain_curve(analysis_object)
}
\arguments{
\item{analysis_object}{Fitted analysis_object with 'fine_tuning()'.}
}
\value{
analysis_object
}
\description{
Plotting Gain Curve
}
