% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting_utils.R
\name{plot_graph_nn}
\alias{plot_graph_nn}
\title{Plot Neural Network Architecture}
\usage{
plot_graph_nn(analysis_object)
}
\arguments{
\item{analysis_object}{Fitted analysis_object with 'fine_tuning()'.}
}
\value{
analysis_object
}
\description{
Plot Neural Network Architecture
}
