% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting_utils.R
\name{plot_lift_curve}
\alias{plot_lift_curve}
\title{Plotting Lift Curve}
\usage{
plot_lift_curve(analysis_object)
}
\arguments{
\item{analysis_object}{Fitted analysis_object with 'fine_tuning()'.}
}
\value{
analysis_object
}
\description{
The \strong{plot_lift_curve()} function plots lift factor as a function of
population percentile when observations are ranked by descending predicted
probability. The lift factor quantifies model's ranking efficiency relative
to random ordering baseline at each population cumulative segment, showing
how much better model selection performs compared to random case selection.
}
\examples{
# Note: For obtaining the lift curve plot the user needs to complete till
# fine_tuning( ) function of the MLwrap pipeline and only with categorical
# outcome.
# See the full pipeline example under plot_calibration_curve()
# Final call signature:
# plot_lift_curve(wrap_object)
}
\seealso{
\code{\link{plot_calibration_curve}}
}
