% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting_utils.R
\name{plot_lift_curve}
\alias{plot_lift_curve}
\title{Plotting Lift Curve}
\usage{
plot_lift_curve(analysis_object)
}
\arguments{
\item{analysis_object}{Fitted analysis_object with 'fine_tuning()'.}
}
\value{
analysis_object
}
\description{
Plotting Lift Curve
}
