% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting_utils.R
\name{plot_loss_curve}
\alias{plot_loss_curve}
\title{Plot Neural Network Loss Curve}
\usage{
plot_loss_curve(analysis_object)
}
\arguments{
\item{analysis_object}{Fitted analysis_object with 'fine_tuning()'.}
}
\value{
analysis_object
}
\description{
Plots the training loss curve of the Neural Network model on the validation
set. This plot can be used for underfitting / overfitting diagnostics.
}
\examples{

# Note: For obtaining the loss curve plot the user needs to
# complete till the fine_tuning( ) function of the MLwrap pipeline.

  # See the full pipeline example under table_best_hyperparameters()
  # (Neural Network engine required)
  # Final call signature:
  # plot_loss_curve(wrap_object)

}
\seealso{
\code{\link{table_best_hyperparameters}}
}
