% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting_utils.R
\name{plot_loss_curve}
\alias{plot_loss_curve}
\title{Plot Neural Network Loss Curve}
\usage{
plot_loss_curve(analysis_object)
}
\arguments{
\item{analysis_object}{Fitted analysis_object with 'fine_tuning()'.}
}
\value{
analysis_object
}
\description{
Plot Neural Network Loss Curve
}
