% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting_utils.R
\name{plot_olden}
\alias{plot_olden}
\title{Plotting Olden Values Barplot}
\usage{
plot_olden(analysis_object, show_table = FALSE)
}
\arguments{
\item{analysis_object}{Fitted analysis_object with 'sensitivity_analysis(methods = "Olden")'.}

\item{show_table}{Boolean. Whether to print Olden results table.}
}
\value{
analysis_object
}
\description{
Plotting Olden Values Barplot
}
