% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting_utils.R
\name{plot_pfi}
\alias{plot_pfi}
\title{Plotting Permutation Feature Importance Barplot}
\usage{
plot_pfi(analysis_object, show_table = FALSE)
}
\arguments{
\item{analysis_object}{Fitted analysis_object with 'sensitivity_analysis(methods = "PFI")'.}

\item{show_table}{Boolean. Whether to print PFI results table.}
}
\value{
analysis_object
}
\description{
Plotting Permutation Feature Importance Barplot
}
