% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting_utils.R
\name{plot_residuals_distribution}
\alias{plot_residuals_distribution}
\title{Plotting Residuals Distribution}
\usage{
plot_residuals_distribution(analysis_object)
}
\arguments{
\item{analysis_object}{Fitted analysis_object with 'fine_tuning()'.}
}
\value{
analysis_object
}
\description{
Plotting Residuals Distribution
}
