% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting_utils.R
\name{plot_roc_curve}
\alias{plot_roc_curve}
\title{Plotting ROC Curve}
\usage{
plot_roc_curve(analysis_object)
}
\arguments{
\item{analysis_object}{Fitted analysis_object with 'fine_tuning()'.}
}
\value{
analysis_object
}
\description{
Plotting ROC Curve
}
