% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting_utils.R
\name{plot_scatter_predictions}
\alias{plot_scatter_predictions}
\title{Plotting Observed vs Predictions}
\usage{
plot_scatter_predictions(analysis_object)
}
\arguments{
\item{analysis_object}{Fitted analysis_object with 'fine_tuning()'.}
}
\value{
analysis_object
}
\description{
The \strong{plot_scatter_predictions()} function generates scatter plots between
observed and predicted values, providing direct visual assessment of model
predictive accuracy. The function displays both training and test results side
by side, enabling evaluation of model generalization performance and
identification of potential overfitting.
}
\examples{
# Note: For obtaining the observed vs. predicted values plot the user needs to
# complete till fine_tuning( ) function of the MLwrap pipeline.

# See the full pipeline example under table_best_hyperparameters()
# Final call signature:
# plot_scatter_predictions(wrap_object)

}
\seealso{
\code{\link{table_best_hyperparameters}}
}
