% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting_utils.R
\name{plot_shap}
\alias{plot_shap}
\title{Plotting SHAP Plots}
\usage{
plot_shap(analysis_object, show_table = FALSE)
}
\arguments{
\item{analysis_object}{Fitted analysis_object with 'sensitivity_analysis(methods = "SHAP")'.}

\item{show_table}{Boolean. Whether to print SHAP summarized results table.}
}
\value{
analysis_object
}
\description{
Mean Abs Barplot, Directional Barplot, Boxplot, Swarmplot
}
