% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting_utils.R
\name{plot_sobol_jansen}
\alias{plot_sobol_jansen}
\title{Plotting Sobol-Jansen Values Barplot}
\usage{
plot_sobol_jansen(analysis_object, show_table = FALSE)
}
\arguments{
\item{analysis_object}{Fitted analysis_object with 'sensitivity_analysis(methods = "Sobol_Jansen")'.}

\item{show_table}{Boolean. Whether to print Sobol-Jansen results table.}
}
\value{
analysis_object
}
\description{
Plotting Sobol-Jansen Values Barplot
}
