% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting_utils.R
\name{plot_tuning_results}
\alias{plot_tuning_results}
\title{Plotting Tuner Search Results}
\usage{
plot_tuning_results(analysis_object)
}
\arguments{
\item{analysis_object}{Fitted analysis_object with 'fine_tuning()'.}
}
\value{
analysis_object
}
\description{
Plotting Tuner Search Results
}
