% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/InterpretableML.R
\name{sensitivity_analysis}
\alias{sensitivity_analysis}
\title{Perform Sensitivity Analysis and Interpretable ML methods}
\usage{
sensitivity_analysis(analysis_object, methods = c("PFI"), metric = NULL)
}
\arguments{
\item{analysis_object}{analysis_object created from fine_tuning function.}

\item{methods}{Method to be used. A string of the method name: "PFI" (Permutation Feature Importance),
"SHAP" (SHapley Additive exPlanations), "Integrated Gradients" (Neural Network only) or
"Olden" (Neural Network only).}

\item{metric}{Metric used for "PFI" method (Permutation Feature Importance).
A string of the name of metric (see Metrics).}
}
\value{
Updated analysis_object
}
\description{
Perform Sensitivity Analysis and Interpretable ML methods
}
