% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table_utils.R
\name{table_evaluation_results}
\alias{table_evaluation_results}
\title{Evaluation Results}
\usage{
table_evaluation_results(analysis_object, show_table = FALSE)
}
\arguments{
\item{analysis_object}{Fitted analysis_object with 'fine_tuning()'.}

\item{show_table}{Boolean. Whether to print the table.}
}
\value{
Tibble or list of tibbles (multiclass classification) with evaluation results.
}
\description{
Evaluation Results
}
