% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table_utils.R
\name{table_integrated_gradients_results}
\alias{table_integrated_gradients_results}
\title{Integrated Gradients Summarized Results Table}
\usage{
table_integrated_gradients_results(analysis_object, show_table = FALSE)
}
\arguments{
\item{analysis_object}{Fitted analysis_object with 'sensitivity_analysis(methods = "Integrated Gradients")'.}

\item{show_table}{Boolean. Whether to print the table.}
}
\value{
Tibble or list of tibbles (multiclass classification) with Integrated Gradient summarized results.
}
\description{
To summarize the Integrated Gradients values calculated, three different metrics are computed:
\itemize{
\item \strong{Mean Absolute Value}
\item \strong{Standard Deviation of Mean Absolute Value}
\item \strong{Directional Sensitivity Value} (Cov(Feature values, IG values) / Var(Feature values))
}
}
