% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table_utils.R
\name{table_sobol_jansen_results}
\alias{table_sobol_jansen_results}
\title{Sobol-Jansen Results Table}
\usage{
table_sobol_jansen_results(analysis_object, show_table = FALSE)
}
\arguments{
\item{analysis_object}{Fitted analysis_object with 'sensitivity_analysis(methods = "Sobol_Jansen")'.}

\item{show_table}{Boolean. Whether to print the table.}
}
\value{
Tibble with Sobol-Jansen results.
}
\description{
Sobol-Jansen Results Table
}
